oh-tab-ycg2: extract LLM client/config helpers
